Change `Browser` test selectors to component name
The selectors for Enzyme are like jQuery selectors. But here, we can
specify the `Component.displayName`. Babel, or React, or someone seems
to automatically set those equal to the class name when it is not given.
So [if names start with a capital
letter](http://airbnb.io/enzyme/docs/api/selector.html#3-a-react-component-s-displayname)
we can query based on those. Seems better than class names, wich may
need to be changed for styling purposes in the future.